requirements.txt
agno
openai
python-dotenv
duckduckgo_search
lancedb 
tantivy 
pypdf 
streamlit
yfinance
sqlalchemy
fastapi
uvicorn